unproxy

+++ packages/packdb/deepProxy.js
@@ -1,9 +1,11 @@
 function createDeepProxy(target, handler) {
+    const preproxy = new WeakMap();
+    
     function makeHandler(path) {
         return {
             set(target, key, value, receiver) {
                 if(typeof value === 'object') {
-                    value = makeProxy(value, [...path, key]);
+                    value = proxify(value, [...path, key]);
                 }
                 target[key] = value;
 
@@ -15,11 +17,10 @@ function createDeepProxy(target, handler) {
 
             deleteProperty(target, key) {
                 if(Reflect.has(target, key)) {
+                    unproxy(target, key);
                     let deleted = Reflect.deleteProperty(target, key);
-                    if(deleted) {
-                        if(handler.deleteProperty) {
-                            handler.deleteProperty(target, [...path, key]);
-                        }
+                    if(deleted && handler.deleteProperty) {
+                        handler.deleteProperty(target, [...path, key]);
                     }
                     return deleted;
                 }
@@ -28,16 +29,33 @@ function createDeepProxy(target, handler) {
         }
     }
     
-    function makeProxy(obj, path) {
-        for(let key of Object.keys(obj)) {
-            if(typeof obj[key] === 'object') {
-                obj[key] = makeProxy(obj[key], [...path, key]);
+    function unproxy(obj, key) {
+        if(preproxy.has(obj[key])) {
+            // console.log('unproxy',key);
+            obj[key] = preproxy.get(obj[key]);
+            // preproxy.delete(obj[key]);
+        }
+        
+        for(let k of Object.keys(obj[key])) {
+            if(typeof obj[key][k] === 'object') {
+                unproxy(obj[key], k);
             }
         }
-        return new Proxy(obj, makeHandler(path))
+
     }
     
-    return makeProxy(target, []);
+    function proxify(obj, path) {
+        for(let key of Object.keys(obj)) {
+            if(typeof obj[key] === 'object') {
+                obj[key] = proxify(obj[key], [...path, key]);
+            }
+        }
+        let p = new Proxy(obj, makeHandler(path));
+        preproxy.set(p, obj);
+        return p;
+    }
+    
+    return proxify(target, []);
 }
